+++ strict.js
@@ -152,8 +152,7 @@ module.exports = [
 
 			// enact plugin https://github.com/enactjs/eslint-plugin-enact/
 			'enact/display-name': 'warn',
-			'enact/kind-name': 'warn',
-			'enact/prop-types': ['warn', {'ignore': ['children', 'className', 'style']}]
+			'enact/kind-name': 'warn'
 		}
 	},
 	{
@@ -193,9 +192,7 @@ module.exports = [
 			}, {
 				"object": "it",
 				"property": "only"
-			}],
-			// PropType validation is not a priority in tests
-			'enact/prop-types': 'off'
+			}]
 		}
 	}
 ];